feat: 🎸 rename modules to module instead of methods
