+++ src/demo/multicore-markdown/module.ts
