+++ src/demo/multicore-math/info.ts
@@ -1,7 +1,7 @@
 import {resolve} from 'path';
 
-export type {Methods} from './methods';
+export type {Methods} from './module';
 
 export const id = 'math';
 
-export const file = resolve(__dirname, 'methods');
+export const file = resolve(__dirname, 'module');
