+++ src/demo/multicore-sleep/module.ts
