+++ src/demo/multicore-tests/info.ts
@@ -1,7 +1,7 @@
 import {resolve} from 'path';
 
-export type {Methods} from './methods';
+export type {Methods} from './module';
 
 export const id = 'tests';
 
-export const file = resolve(__dirname, 'methods');
+export const file = resolve(__dirname, 'module');
